requirements.txt
langchain
streamlit
langchain_google_genai
chromadb
google-generativeai
langchain-community
langchain-experimental